02_plant_disease_classification/requirements_api.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
pandas>=2.0.0
python-multipart>=0.0.9
requests>=2.31.0
